requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
discord.py==2.3.2
httpx==0.27.0
pydantic==2.6.4
python-dotenv==1.0.1
